little more refactoring
