+++ SysKit.ODG.App/SysKit.ODG.Base/Interfaces/Generation/IGenerationService.cs
@@ -3,6 +3,7 @@ using System.Collections.Generic;
 using System.Text;
 using System.Threading.Tasks;
 using SysKit.ODG.Base.DTO.Generation;
+using SysKit.ODG.Base.DTO.Generation.Options;
 using SysKit.ODG.Base.Interfaces.Authentication;
 
 namespace SysKit.ODG.Base.Interfaces.Generation
