+++ SysKit.ODG.App/SysKit.ODG.Base/Interfaces/Generation/IGenerationTask.cs
@@ -3,6 +3,7 @@ using System.Collections.Generic;
 using System.Text;
 using System.Threading.Tasks;
 using SysKit.ODG.Base.DTO.Generation;
+using SysKit.ODG.Base.DTO.Generation.Options;
 
 namespace SysKit.ODG.Base.Interfaces.Generation
 {
