+++ SysKit.ODG.App/SysKit.ODG.Base/Interfaces/Generation/IUserDataGeneration.cs
@@ -2,6 +2,7 @@
 using System.Collections.Generic;
 using System.Text;
 using SysKit.ODG.Base.DTO.Generation;
+using SysKit.ODG.Base.DTO.Generation.Options;
 
 namespace SysKit.ODG.Base.Interfaces.Generation
 {
